+++ src/pretalx/common/mixins/__init__.py
@@ -1,5 +1,5 @@
 from .forms import ReadOnlyFlag
-from .models import LogMixin
+from .models import LogMixin, IdBasedSlug
 from .views import ActionFromUrl, Filterable, PermissionRequired, Sortable
 
 __all__ = [