+++ doc/changelog.rst
@@ -3,6 +3,7 @@
 Release Notes
 =============
 
+- :feature:`628` Prefill CfP form with track and submission type if they are available in the query string.
 - :support:`-` pretalx is now able to run with Python 3.7 even when using Celery.
 - :feature:`598` Via the new review phase settings, speakers can be allowed to modify their submissions while they are being reviewed.
 - :feature:`641` At the undocumented URL `/<event>/speaker/by-id/123/` there is now a redirect to the canonical speaker URL `/<event>/speaker/CODE/` to work around export format restrictions.